requirements.txt
drf-yasg==1.21.8
djangorestframework-simplejwt==5.3.1
djangorestframework==3.15.2
psycopg2-binary
django-environ